Make utimes work under Linux as well
It seems that utimes will completely ignore the call
under Linux when btime is set. Therefore I have split
it up into two separate calls of utimes.

It also does not throw an error if the file does not exist,
so I adapted a test because I don't think this is required.